+++ client/src/Components/Pages/Profile/index.js
@@ -1,6 +1,5 @@
 import React, { Component } from "react";
 import axios from "axios";
-import moment from "moment";
 import { message, Skeleton } from "antd";
 
 // import MonthlyReviews from "./ProfileAnswers/MonthlyReviews";
@@ -13,6 +12,8 @@ import DetailedSection from "./DetailedSection";
 
 import Layout from "../../Common/Layout";
 import { Wrapper } from "./Profile.style";
+
+import { getContractorsFromReviews } from "./utils";
 
 export default class Profile extends Component {
   state = {
@@ -70,18 +71,7 @@ export default class Profile extends Component {
           summary[0].category === "worksite" &&
           reviewDetails.length
         ) {
-          const [worksiteQuestionsGroup] = reviewDetails.filter(
-            group => group._id === "Working on the site"
-          );
-          const [contractorQuestion] = worksiteQuestionsGroup.questions.filter(
-            question => question.text === "Who was the main contractor on site?"
-          );
-          // question => question.text === "Who is the main contractor on site?"
-          const orderedAnswers = contractorQuestion.answers.sort(
-            (a, b) =>
-              moment(a.updatedAt).valueOf() - moment(b.updatedAt).valueOf()
-          );
-          contractorAnswers = orderedAnswers.map(item => item.answer);
+          contractorAnswers = getContractorsFromReviews(reviewDetails);
         }
 
         this.setState({
@@ -100,37 +90,6 @@ export default class Profile extends Component {
           err.response && err.response.data && err.response.data.error;
         message.error(error || "Something went wrong");
       });
-  };
-
-  getCarCost = () => {
-    const { reviewDetails } = this.state;
-
-    // get the car parking cost question
-    const carSection = reviewDetails
-      .filter(section => section._id === null)
-      .map(item =>
-        item.questions.filter(
-          question => question.text === "How much did car parking cost per day?"
-        )
-      );
-
-    if (!carSection || carSection.length < 1) return "N/A";
-
-    // work out the average cost from the answers
-    const costsArr = carSection[0][0].answers.map(answer => answer.answer);
-
-    const average =
-      costsArr.reduce((accum, curr) => {
-        return accum + curr;
-      }, 0) / costsArr.length;
-
-    if (average > 0) {
-      if (Number.isInteger(average)) {
-        return average;
-      }
-      return average.toFixed(2);
-    }
-    return "Free";
   };
 
   updateLastViewed = () => {
@@ -202,37 +161,6 @@ export default class Profile extends Component {
     }
   };
 
-  reviewsByMonth = () => {
-    const { FilteredReviewMonths } = this.state;
-
-    const reviewMonths = FilteredReviewMonths.map(review => {
-      return moment(review.createdAt).format("MMM");
-    });
-
-    const reviewMonthsCount = {
-      Jan: 0,
-      Feb: 0,
-      Mar: 0,
-      Apr: 0,
-      May: 0,
-      Jun: 0,
-      Jul: 0,
-      Aug: 0,
-      Sep: 0,
-      Oct: 0,
-      Nov: 0,
-      Dec: 0,
-    };
-
-    if (FilteredReviewMonths.length === 0) return reviewMonthsCount;
-
-    reviewMonths.forEach(month => {
-      reviewMonthsCount[month] += 1;
-    });
-
-    return reviewMonthsCount;
-  };
-
   render() {
     const {
       summary,
@@ -264,10 +192,6 @@ export default class Profile extends Component {
     } = this.props;
 
     // if (!loaded) return <Loading />;
-
-    if (!loaded) {
-      return null;
-    }
 
     return (
       <Layout type="center">
@@ -318,7 +242,6 @@ export default class Profile extends Component {
               reviewDetails={reviewDetails}
               summary={summary}
               toggleComments={this.toggleComments}
-              getCarCost={this.getCarCost}
             />
           )}
